Rename to Notification and remove UpdateMessageDto
