+++ src/test/controllers/notifications.controller.ts
@@ -1,22 +1,19 @@
 import { Controller, Get, HttpService, Post, Body } from "@nestjs/common";
 import { ApiTags, ApiOperation } from "@nestjs/swagger";
-import { UpdateMessage } from "../../shared/dto/update-message.dto";
-import { MESSAGES_ALL } from "../../../test/mocks/update-messages.mock";
-import { UpdateMessageTestDto } from "../update-test-message.dto";
 
 @ApiTags("test")
 @Controller("notifications")
 export class NotificationsController {
 
 	constructor(private http: HttpService) { }
-
+	// TODO: Add new examples
 	@Post()
 	@ApiOperation({
 		operationId: "sendNotification",
 		summary: "Send notication.",
 		description: "Sends the given UpdateMessage to the specified URL via HTTP-POST. Use (GET) /notifications for examples."
 	})
-	sendNotification(@Body() testMessage: UpdateMessageTestDto): void {
+	sendNotification(@Body() testMessage: any): void {
 		console.log("Sending message to: " + testMessage.url);
 		this.http.post(testMessage.url, testMessage.message).toPromise();
 	} 
@@ -27,7 +24,7 @@ export class NotificationsController {
 		summary: "SHOULD NOT BE CALLED.",
 		description: "Logs the received UpdateMessage send from (POST) /notifications, if it was send to this route (/notifications/log)."
 	})
-	logReceivedMessage(@Body() message: UpdateMessage): void {
+	logReceivedMessage(@Body() message: any): void {
 		console.log(message);
 	}
 
@@ -37,9 +34,9 @@ export class NotificationsController {
 		summary: "Get UpdateMessage-Examples.",
 		description: "Returns all possible UpdateMessages."
 	})
-	getNotification(): UpdateMessage[] {
+	getNotification(): any[] {
 		console.log("getNotifactions() was called");
-		return MESSAGES_ALL;
+		return null;
 	}
 	
 }
